Logo's and title update. (#269)
* The logo's in app attack has been setup.

In the App attack project the logo's were missing and I have uploaded the logo's for the missing ones.

* The Title has been updated to the correct one.

The Heading in the Deakin Cyber safe VR project is changed from app attack to the correct project heading. But still the entire page is made for the app attack project and not the cybersafe VR project.